get current user created
